BackendServices/requirements.txt
fastapi
uvicorn
python-docx
pymupdf
sentence-transformers
faiss-cpu
openai
dotenv
python-multipart
pandas